arm64: dts: qcom: Disable CoreSight DT entries for SM8150
CoreSight is a set of trace and debug software by Arm Limited. and is used
on their architecture of SoCs and is very instrumental in early bring-up or
in development stage of boards. These tracing features are useful in such
scenarios but it comes with its heavy overhead.

CoreSight tracing configs are enabled by default on perf-defconfigs but is
recommended to be disabled on production devices as per 80-PJ035-1 article
due its heavy overhead, even when these debugging configs are disabled due
to these DT entries clocks are being left on which in turn introduces power
regressions.

This commit marks JTAG driver nodes as disabled and removes the inclusion of
the coresight devicetree with its properties and mappings, also remove other
nodes that were reduntant due to this change.

Signed-off-by: Daniel Jacob Chittoor <djchittoor47@gmail.com>
(cherry picked from commit 2b489b04aecfac18df8b85300cda95d2120e01f5)